Update README.md
Added qbee.io -  we do docker orchestration, reverse proxy, are nearly a caas platform with focus on IoT as well as servers.
We are not open source, but have a free tier with 2 devices. Not sure if we fit here, but being able to join this list would be really appreciated.